Fix TPC-H partitioned schema to use BIGINT (for > 1Tb loads)

+++ ddl-tpch/bin_partitioned/orders.sql
@@ -3,7 +3,7 @@ use ${DB};
 
 drop table if exists orders;
 
-create table orders (O_ORDERKEY INT,
+create table orders (O_ORDERKEY BIGINT,
  O_CUSTKEY BIGINT,
  O_ORDERSTATUS STRING,
  O_TOTALPRICE DOUBLE,
